improved float / double handling

+++ ai/coreai.cpp
@@ -2040,18 +2040,18 @@ void CoreAI::sortUnitsFarFromEnemyFirst(std::vector<MoveUnitData> & pUnits, spQm
 bool CoreAI::needsRefuel(const Unit *pUnit) const
 {
     if (pUnit->getMaxFuel() > 0 &&
-        pUnit->getFuel() / static_cast<float>(pUnit->getMaxFuel()) < m_fuelResupply)
+        static_cast<float>(pUnit->getFuel()) / static_cast<float>(pUnit->getMaxFuel()) < m_fuelResupply)
     {
         return true;
     }
     if (pUnit->getMaxAmmo1() > 0 &&
-        pUnit->getAmmo1() / static_cast<float>(pUnit->getMaxAmmo1()) < m_ammoResupply &&
+        static_cast<float>(pUnit->getAmmo1()) / static_cast<float>(pUnit->getMaxAmmo1()) < m_ammoResupply &&
         !pUnit->getWeapon1ID().isEmpty())
     {
         return true;
     }
     if (pUnit->getMaxAmmo2() > 0 &&
-        pUnit->getAmmo2() / static_cast<float>(pUnit->getMaxAmmo2()) < m_ammoResupply &&
+        static_cast<float>(pUnit->getAmmo2()) / static_cast<float>(pUnit->getMaxAmmo2()) < m_ammoResupply &&
         !pUnit->getWeapon2ID().isEmpty())
     {
         return true;
